move mocks and tests out of lib
