+++ packages/ui-components/src/__mocks__/MockListViewOrderbookFilters.svelte
@@ -1,8 +1,17 @@
 <script lang="ts">
+	export let query: any;
+	export let activeSubgraphs: any;
+	export let settings: any;
+	export let accounts: any;
+	export let activeAccountsItems: any;
+	export let activeOrderStatus: any;
+	export let orderHash: any;
+	export let hideZeroBalanceVaults: any;
 	export let isVaultsPage = false;
 	export let isOrdersPage = false;
 </script>
 
+/** eslint-disable svelte/valid-compile */ /** eslint-disable svelte/valid-compile */
 <div data-testid="orderbook-filters">
 	{#if isVaultsPage}
 		<div data-testid="zero-balance-filter"></div>
