+++ packages/ui-components/src/__tests__/DropdownActiveSubgraphs.test.ts
@@ -1,8 +1,8 @@
 import { render, fireEvent, screen, waitFor } from '@testing-library/svelte';
 import { get, writable, type Writable } from 'svelte/store';
 import { beforeEach, expect, test, describe } from 'vitest';
-import DropdownActiveSubgraphs from './DropdownActiveSubgraphs.svelte';
-import { mockConfigSource } from '../../__mocks__/settings';
+import DropdownActiveSubgraphs from '../lib/components/dropdown/DropdownActiveSubgraphs.svelte';
+import { mockConfigSource } from '../__mocks__/settings';
 
 describe('DropdownActiveSubgraphs', () => {
 	const mockSettings = {
