+++ packages/ui-components/src/__tests__/Hash.test.ts
@@ -1,5 +1,5 @@
 import { render } from '@testing-library/svelte';
-import Hash from './Hash.svelte';
+import Hash from '../lib/components/Hash.svelte';
 import { describe, it, expect } from 'vitest';
 import userEvent from '@testing-library/user-event';
