+++ packages/ui-components/src/__tests__/ListViewOrderbookFilters.test.ts
@@ -1,8 +1,8 @@
 import { render, screen } from '@testing-library/svelte';
 import { writable } from 'svelte/store';
 import { beforeEach, expect, test, describe } from 'vitest';
-import ListViewOrderbookFilters from './ListViewOrderbookFilters.svelte';
-import type { ConfigSource } from '../typeshare/config';
+import ListViewOrderbookFilters from '../lib/components/ListViewOrderbookFilters.svelte';
+import type { ConfigSource } from '../lib/typeshare/config';
 import userEvent from '@testing-library/user-event';
 import { createResolvableInfiniteQuery } from '../__mocks__/queries';
 import type { CreateInfiniteQueryResult, InfiniteData } from '@tanstack/svelte-query';
