+++ packages/ui-components/src/__tests__/OrdersListTable.test.ts
@@ -1,5 +1,5 @@
 import { render, screen } from '@testing-library/svelte';
-import OrdersListTable from './OrdersListTable.svelte';
+import OrdersListTable from '../lib/components/tables/OrdersListTable.svelte';
 import type { ComponentProps } from 'svelte';
 
 const {
@@ -10,7 +10,7 @@ const {
 	mockActiveOrderStatusStore,
 	mockOrderHashStore,
 	mockHideZeroBalanceVaultsStore
-} = await vi.hoisted(() => import('../../__mocks__/stores'));
+} = await vi.hoisted(() => import('../__mocks__/stores'));
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 type OrdersListTableProps = ComponentProps<OrdersListTable<any>>;
@@ -24,13 +24,13 @@ vi.mock('@tanstack/svelte-query', async (importOriginal) => {
 
 vi.mock('../ListViewOrderbookFilters.svelte', async () => {
 	return {
-		default: (await import('../../__mocks__/MockListViewOrderbookFilters.svelte')).default
+		default: await import('../__mocks__/MockListViewOrderbookFilters.svelte')
 	};
 });
 
 vi.mock('../TanstackAppTable.svelte', async () => {
 	return {
-		default: (await import('../../__mocks__/MockTanstackAppTable.svelte')).default
+		default: await import('../__mocks__/MockTanstackAppTable.svelte')
 	};
 });
